First pass at structure refactor.
