+++ omni_blocks/blocks/__init__.py
@@ -1,0 +1,5 @@
+from omni_blocks.blocks.body import *  # noqa
+from omni_blocks.blocks.chooser_blocks import *  # noqa
+from omni_blocks.blocks.image_blocks import *  # noqa
+from omni_blocks.blocks.list_blocks import *  # noqa
+from omni_blocks.blocks.struct_blocks import *  # noqa
